Fixing PR remarks (#94)
