+++ setup.cfg
@@ -23,4 +23,3 @@ skip = httpcore/_sync/,tests/sync_tests/
 addopts = --cov-report= --cov=httpcore --cov=tests -rxXs
 markers =
   copied_from(source, changes=None): mark test as copied from somewhere else, along with a description of changes made to accodomate e.g. our test setup
-  curio: mark the test as a coroutine, it will be run using a Curio kernel.
