+++ tests/conftest.py
@@ -1,16 +1,11 @@
 import asyncio
 import contextlib
-import functools
-import inspect
 import os
 import ssl
 import threading
 import time
 import typing
 
-import curio
-import curio.debug
-import curio.meta
 import pytest
 import trustme
 import uvicorn
@@ -19,53 +14,12 @@ from mitmproxy.tools.dump import DumpMaster
 
 from httpcore._types import URL
 
+from .marks.curio import kernel  # noqa: F401
+from .marks.curio import pytest_pycollect_makeitem  # noqa: F401
+from .marks.curio import pytest_pyfunc_call  # noqa: F401
+
 PROXY_HOST = "127.0.0.1"
 PROXY_PORT = 8080
-
-
-def _is_coroutine(obj):
-    """Check to see if an object is really a coroutine."""
-    return curio.meta.iscoroutinefunction(obj) or inspect.isgeneratorfunction(obj)
-
-
-@pytest.mark.tryfirst
-def pytest_pycollect_makeitem(collector, name, obj):
-    """A pytest hook to collect coroutines in a test module."""
-    if collector.funcnamefilter(name) and _is_coroutine(obj):
-        item = pytest.Function.from_parent(collector, name=name)
-        if "curio" in item.keywords:
-            return list(collector._genfunctions(name, obj))
-
-
-@pytest.hookimpl(tryfirst=True, hookwrapper=True)
-def pytest_pyfunc_call(pyfuncitem):
-    """Run curio marked test functions in a Curio kernel instead of a normal function call.
-    """
-    if pyfuncitem.get_closest_marker("curio"):
-        pyfuncitem.obj = wrap_in_sync(pyfuncitem.obj)
-    yield
-
-
-def wrap_in_sync(func):
-    """Return a sync wrapper around an async function executing it in a Kernel."""
-
-    @functools.wraps(func)
-    def inner(**kwargs):
-        coro = func(**kwargs)
-        curio.Kernel().run(coro, shutdown=True)
-
-    return inner
-
-
-# Fixture for explicitly running in Kernel instance.
-@pytest.fixture(scope="session")
-def kernel(request):
-    """Provide a Curio Kernel object for running co-routines."""
-    k = curio.Kernel(debug=[curio.debug.longblock, curio.debug.logcrash])
-    m = curio.monitor.Monitor(k)
-    request.addfinalizer(lambda: k.run(shutdown=True))
-    request.addfinalizer(m.close)
-    return k
 
 
 @pytest.fixture(
